requirements.txt
altair==5.1.2
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
boto3==1.28.71
botocore==1.31.71
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
fonttools==4.43.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.5.0
jupyter_core==5.4.0
kiwisolver==1.4.5
lightgbm==4.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
s3transfer==0.7.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.28.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.12.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.8
xgboost==2.0.1
zipp==3.17.0
